Cast to Float, not Integer

+++ tcms/rpc/api/testcase.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from django.db.models import IntegerField
+from django.db.models import FloatField
 from django.db.models.functions import Cast, Coalesce
 from django.forms import EmailField, ValidationError
 from django.forms.models import model_to_dict
@@ -290,9 +290,9 @@ def filter(query=None):  # pylint: disable=redefined-builtin
     # `Cast`ed to their millisecond values
     qs = (
         TestCase.objects.annotate(
-            setup_duration_sec=Cast("setup_duration", IntegerField()) / 10 ** 6,
-            testing_duration_sec=Cast("testing_duration", IntegerField()) / 10 ** 6,
-            expected_duration_sec=Coalesce("setup_duration_sec", 0) + Coalesce("testing_duration_sec", 0),
+            setup_duration_sec=Cast("setup_duration", FloatField()) / 10 ** 6,
+            testing_duration_sec=Cast("testing_duration", FloatField()) / 10 ** 6,
+            expected_duration_sec=Coalesce("setup_duration_sec", 0.0) + Coalesce("testing_duration_sec", 0.0),
         )
         .filter(**query)
         .values(
